Set the board speed, add objects at an interval, and introduce gameOver
We now have a new property on the board which we can later set with
a slider to change the overall speed at which the objects fall.

We also continually add a new object to the canvas at a current set
interval of one second.

Finally, we stop the game, reset the score, and stop the generation of
objects.