Begin styling and making app responsive
Create a div for header elements which include the game title, score,
start button and speed slider.

Add a <meta> to set viewport.

Create a listener on the window to check and resize the canvas if the
window is scaled up/down.

Set some global variables on canvas height and width. update canvas
and beginning x position of objects with new width.
Reset the canvas width if the screen is resized.